py_test/requirement.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
et-xmlfile==1.1.0
execnet==1.9.0
iniconfig==1.1.1
numpy==1.22.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
toml==0.10.2
